Add RDKit fragment comparison and refactor pipeline
Added compare_rdkit_fragments.py for RDKit-based fragment comparison utilities. Refactored run_reaction_template_pipeline.py to only define reaction dictionaries and moved the main reaction processing logic to a new run_reaction_template_pipeline_2.py, which now contains the reaction_selector function and main execution block.
